+++ com.unity.render-pipelines.lightweight/Runtime/LightweightRenderPipeline.cs
@@ -11,6 +11,11 @@ using Lightmapping = UnityEngine.Experimental.GlobalIllumination.Lightmapping;
 
 namespace UnityEngine.Rendering.LWRP
 {
+    public interface IBeforeCameraRender
+    {
+        void ExecuteBeforeCameraRender(LightweightRenderPipeline pipelineInstance, ScriptableRenderContext context, Camera camera);
+    }
+
     public sealed partial class LightweightRenderPipeline : UnityEngine.Rendering.RenderPipeline
     {
         static class PerFrameBuffer
@@ -165,13 +170,16 @@ namespace UnityEngine.Rendering.LWRP
         {
             BeginFrameRendering(cameras);
 
-            GraphicsSettings.lightsUseLinearIntensity = true;
+            GraphicsSettings.lightsUseLinearIntensity = (QualitySettings.activeColorSpace == ColorSpace.Linear);
             SetupPerFrameShaderConstants();
 
             SortCameras(cameras);
             foreach (Camera camera in cameras)
             {
                 BeginCameraRendering(camera);
+
+                foreach (var beforeCamera in camera.GetComponents<IBeforeCameraRender>())
+                    beforeCamera.ExecuteBeforeCameraRender(this, renderContext, camera);
 
                 RenderSingleCamera(this, renderContext, camera, camera.GetComponent<IRendererSetup>());
             }